Core/Spell: revamp of immunities system

+++ Source/Game/Entities/Vehicle.cs
@@ -151,8 +151,8 @@ namespace Game.Entities
                 _me.ApplySpellImmune(0, SpellImmunity.State, AuraType.SchoolImmunity, true);
                 _me.ApplySpellImmune(0, SpellImmunity.State, AuraType.ModUnattackable, true);
                 _me.ApplySpellImmune(0, SpellImmunity.State, AuraType.SchoolAbsorb, true);
-                _me.ApplySpellImmune(0, SpellImmunity.Mechanic, Mechanics.Shield, true);
-                _me.ApplySpellImmune(0, SpellImmunity.Mechanic, Mechanics.Immune_Shield, true);
+                _me.ApplySpellImmune(0, SpellImmunity.Mechanic, (uint)Mechanics.Shield, true);
+                _me.ApplySpellImmune(0, SpellImmunity.Mechanic, (uint)Mechanics.Immune_Shield, true);
 
                 // ... Resistance, Split damage, Change stats ...
                 _me.ApplySpellImmune(0, SpellImmunity.State, AuraType.DamageShield, true);
